requirements.txt
anyio==4.10.0
base58==2.1.1
blinker==1.9.0
cachetools==4.2.4
certifi==2025.8.3
click==8.2.1
construct==2.10.68
construct-typing==0.5.6
Flask==3.1.2
gunicorn==23.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonalias==0.1.1
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.1.1
rfc3986==1.5.0
setuptools==80.9.0
sniffio==1.3.1
solana==0.28.0
solders==0.10.0
types-cachetools==4.2.10
typing_extensions==4.15.0
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
